Automatic precondition update on name change
When a question or question-option's name is changed the UI warns if
there are any preconditions referring to that question.  If there are
and the user accepts then all preconditions will be changed along with
the question or question-option's name.

Added new method: database\qnaire::update_name_in_precondition()
Added new columns to question and question_option modules:
  module_precondition_dependencies
  page_precondition_dependencies
  question_precondition_dependencies
  question_option_precondition_dependencies