Bug: making reverse compatible with old versions of Firefox
As it turns out Firefox didn't support the "s" RegExp flag until 2020,
but since we need to support versions dating back to 2017 the codehas
been changed to do multi-line . matching without using the "s" flag.
Specifically, the string's newlines are converted to string, the RegExp
is run, then the newlines are converted back.